fix: disallow DocType declaration to prevent XML entity attacks
Also added a couple of configurations to have secure processing of XML based on:
 - OWASP best practices
 - The whitepaper – XML Schema, DTD, and Entity Attacks: A Compendium of Known Techniques

Prior to this PR the validations done on the stylesheet and the XML input weren't the same, meaning XML entity attacks were prevented for the XML input but not the stylesheet.

https://gravitee.atlassian.net/browse/APIM-470

BREAKING CHANGE: This policy now has a secure processing by default.
To keep the previous behavior you need to explicitly deactivate the secure processing.
To do so you need to set `policy.xslt.secure-processing: false` in your Gateway's `gravitee.yml`.